requirements.txt
gunicorn==19.7.1
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.26.post2
Django==1.11.13
django-celery==3.2.2
django-cors-headers==2.2.0
future==0.16.0
kombu==3.0.37
pytz==2018.4
vine==1.1.4

